% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_schema.R
\name{get_user_feedback}
\alias{get_user_feedback}
\title{Get user feedback interactively}
\usage{
get_user_feedback(state, verbose)
}
\arguments{
\item{state}{Current conversation state}

\item{verbose}{Show logs}
}
\value{
Updated conversation state
}
\description{
Get user feedback interactively
}
