% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{A value to check for NULL.}

\item{y}{A default value to return if x is NULL.}
}
\value{
x if not NULL, otherwise y.
}
\description{
Returns the left-hand side if it is not NULL, otherwise returns the
right-hand side. This is useful for providing default values.
}
\examples{
NULL \%||\% "default"
"value" \%||\% "default"
}
