% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{ollama_download_model}
\alias{ollama_download_model}
\title{Download a model from Ollama API}
\usage{
ollama_download_model(.model, .ollama_server = "http://localhost:11434")
}
\arguments{
\item{.model}{The name of the model to download}

\item{.ollama_server}{The URL of the Ollama server (default: "http://localhost:11434")}
}
\value{
No return value, called for side effects (downloads the model to
the Ollama server with progress displayed in the console).
}
\description{
This function sends a request to download a specified model from Ollama's
model library with progress tracking.
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
