% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_client.R
\name{ollama_get_version}
\alias{ollama_get_version}
\title{Get Ollama server version information}
\usage{
ollama_get_version(server_url = "http://localhost:11434")
}
\arguments{
\item{server_url}{The Ollama server URL (without /api/chat)}
}
\value{
Version information or NULL if failed
}
\description{
Get Ollama server version information
}
