% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_client.R
\name{parse_max_tokens_from_error}
\alias{parse_max_tokens_from_error}
\title{Parse max_tokens from error message}
\usage{
parse_max_tokens_from_error(error_message)
}
\description{
Parse max_tokens from error message
}
\keyword{internal}
