% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_client.R, R/ollama.R
\name{ollama_download_model}
\alias{ollama_download_model}
\title{Download a model from Ollama}
\usage{
ollama_download_model(.model, .ollama_server = "http://localhost:11434")

ollama_download_model(.model, .ollama_server = "http://localhost:11434")
}
\arguments{
\item{.model}{The name of the model to download}

\item{.ollama_server}{The URL of the Ollama server (default: "http://localhost:11434")}

\item{model}{The model name to download}

\item{server_url}{The Ollama server URL (without /api/chat)}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
This function sends a request to download a specified model from Ollama's
model library with progress tracking.
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
